+++ udata/core/topic/models.py
@@ -84,6 +84,8 @@ class TopicElement(Auditable, Document):
         super().post_save(sender, document, **kwargs)
         if document.topic and document.element and hasattr(document.element, "id"):
             reindex.delay(*as_task_param(document.element))
+        if document.topic:
+            reindex.delay(*as_task_param(document.topic))
 
     @classmethod
     def post_delete(cls, sender, document, **kwargs):
@@ -91,6 +93,8 @@ class TopicElement(Auditable, Document):
         try:
             if document.topic and document.element and hasattr(document.element, "id"):
                 reindex.delay(*as_task_param(document.element))
+            if document.topic:
+                reindex.delay(*as_task_param(document.topic))
         except DoesNotExist:
             # Topic might have been deleted, causing dereferencing to fail
             pass
@@ -114,7 +118,7 @@ class Topic(Datetimed, Auditable, Linkable, Document[OwnedQuerySet], Owned):
     tags = field(ListField(StringField()))
     color = field(IntField())
 
-    featured = field(BooleanField(default=False), auditable=False)
+    featured = field(BooleanField(default=False), readonly=True, auditable=False)
     private = field(BooleanField())
     extras = field(ExtrasField(), auditable=False)